# openapi.yaml
openapi: 3.0.3
info:
  title: AgentX Registry API
  version: 0.8.0
  description: |
    Hosted AgentX API. Business routes require a bearer token and return a stable
    ErrorEnvelope on JSON errors. Unknown dependency failures are logged with the
    request ID and return stable non-sensitive messages. Workspace roles are
    hierarchical: viewer, developer, admin, owner. Legacy unscoped routes are
    disabled in hosted mode. Device and release creation require idempotency keys.
    Expected duplicate and missing-resource outcomes use stable public 4xx envelopes;
    unexpected repository and dependency failures remain non-sensitive 500 responses.
security:
  - bearerAuth: []
servers:
  - url: https://{host}
    description: Hosted deployment
    variables:
      host: { default: registry.example.com }
paths:
...
  /v1/workspaces/{id}/packages:
    get:
      operationId: listWorkspaceReleases
      summary: List immutable package releases
      description: Requires viewer. PostgreSQL applies limit and cursor before returning rows.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/ReleasePage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
  /v1/workspaces/{id}/packages/{name}/releases:
    post:
      operationId: publishWorkspaceRelease
      summary: Publish an immutable Skill release
      description: |
        Requires developer. Artifact is a gzip-compressed POSIX tar archive no larger
        than 51 MiB compressed. It must contain root SKILL.md and only regular
        files/directories. Traversal, duplicate names, links, device entries,
        executable files, credential filenames, and decompression limits are rejected.
        A package/version already present in the workspace, or incompatible reuse of
        an idempotency key, returns a public 400 conflict envelope.
      x-agentx-minimum-role: developer
      parameters:
        - { '$ref': '#/components/parameters/WorkspaceID' }
        - { '$ref': '#/components/parameters/PackageName' }
        - { '$ref': '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [version, artifact]
              properties:
                version: { type: string, minLength: 1 }
                artifact: { type: string, format: binary, description: Deterministic gzip tar Skill archive }
                signature: { type: string, format: byte, description: Base64 Ed25519 signature over exact artifact bytes }
      responses:
        '201':
          description: Published, or replayed with X-Idempotent-Replay true
          headers:
            X-Idempotent-Replay: { schema: { type: string, enum: ['true'] } }
          content:
            application/json: { schema: { '$ref': '#/components/schemas/Release' } }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
  /v1/workspaces/{id}/artifacts/{digest}:
    get:
      operationId: downloadWorkspaceArtifact
      summary: Download an artifact by SHA-256
      description: Requires viewer. Digest must belong to this workspace.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/Digest' }]
      responses:
        '200': { '$ref': '#/components/responses/Artifact' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
  /v1/workspaces/{id}/packages/{name}/{version}/download:
    get:
      operationId: downloadWorkspaceRelease
      summary: Download an approved release
      description: Requires viewer. Pending approval releases are not downloadable.
      x-agentx-minimum-role: viewer
      parameters:
        - { '$ref': '#/components/parameters/WorkspaceID' }
        - { '$ref': '#/components/parameters/PackageName' }
        - { '$ref': '#/components/parameters/Version' }
      responses:
        '200': { '$ref': '#/components/responses/Artifact' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
  /v1/workspaces/{id}/packages/{name}/{version}/approve:
    post:
      operationId: approveWorkspaceRelease
      summary: Approve a pending release
      description: Requires admin.
      x-agentx-minimum-role: admin
      parameters:
        - { '$ref': '#/components/parameters/WorkspaceID' }
        - { '$ref': '#/components/parameters/PackageName' }
        - { '$ref': '#/components/parameters/Version' }
      responses:
        '200': { description: Approved release, content: { application/json: { schema: { '$ref': '#/components/schemas/Release' } } } }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
...
  /v1/workspaces/{id}/devices/{device_id}/plan:
    get:
      operationId: getWorkspaceDevicePlan
      summary: Get the device reconcile plan
      description: Requires viewer.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/DeviceID' }]
      responses:
        '200': { description: Reconcile plan, content: { application/json: { schema: { '$ref': '#/components/schemas/ReconcilePlan' } } } }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
...
  /v1/workspaces/{id}/policies:
    get:
      operationId: getWorkspacePolicy
      summary: Get current policy
      description: Requires viewer.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }]
      responses:
        '200': { description: Current policy, content: { application/json: { schema: { '$ref': '#/components/schemas/Policy' } } } }
        '403': { '$ref': '#/components/responses/Forbidden' }
    put:
      operationId: updateWorkspacePolicy
      summary: Replace policy
      description: Requires admin. Supports require_signature and require_approval.
      x-agentx-minimum-role: admin
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { '$ref': '#/components/schemas/PolicyMutation' } }
      responses:
        '200': { description: New policy revision, content: { application/json: { schema: { '$ref': '#/components/schemas/Policy' } } } }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
...
  /v1/packages:
    get:
      operationId: listLegacyReleases
      deprecated: true
      summary: List releases in local legacy mode
      description: Disabled in hosted mode.
      parameters: [{ '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/ReleasePage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '500': { '$ref': '#/components/responses/InternalError' }
  /v1/packages/{name}/releases:
    post:
      operationId: publishLegacyRelease
      deprecated: true
      summary: Publish in local legacy mode
      description: Disabled in hosted mode.
      parameters: [{ '$ref': '#/components/parameters/PackageName' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [version, artifact]
              properties:
                version: { type: string }
                artifact: { type: string, format: binary }
                signature: { type: string, format: byte }
      responses:
        '201': { description: Published }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '500': { '$ref': '#/components/responses/InternalError' }
  /v1/artifacts/{digest}:
    get:
      operationId: downloadLegacyArtifact
      deprecated: true
      summary: Download in local legacy mode
      description: Disabled in hosted mode.
      parameters: [{ '$ref': '#/components/parameters/Digest' }]
      responses:
        '200': { '$ref': '#/components/responses/Artifact' }
        '404': { '$ref': '#/components/responses/NotFound' }
...
  /v1/drift:
    get:
      operationId: listLegacyDrift
      deprecated: true
      summary: List drift in local legacy mode
      description: Disabled in hosted mode.
      parameters: [{ '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/DriftPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '500': { '$ref': '#/components/responses/InternalError' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: OIDC JWT }
  parameters:
    WorkspaceID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    DeviceID: { name: device_id, in: path, required: true, description: Workspace-scoped opaque text identifier, schema: { type: string, minLength: 1 } }
    PackageName: { name: name, in: path, required: true, schema: { type: string, minLength: 1 } }
    Version: { name: version, in: path, required: true, schema: { type: string, minLength: 1 } }
    Digest: { name: digest, in: path, required: true, schema: { type: string, pattern: '^[a-f0-9]{64}$' } }
    Limit: { name: limit, in: query, required: false, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
    Cursor: { name: cursor, in: query, required: false, description: Opaque value from next_cursor or X-Next-Cursor, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 255 } }
...
  responses:
    Artifact:
      description: Immutable gzip tar artifact bytes
      headers: { Cache-Control: { schema: { type: string } } }
      content:
        application/octet-stream: { schema: { type: string, format: binary } }
...
    ReleasePage: { description: Release page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/Release' } } } }] } } } }
...
    DriftPage: { description: Drift page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/DriftItem' } } } }] } } } }
    BadRequest: { description: Invalid request, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Workspace membership or role denied, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    InternalError: { description: Request failed without exposing dependency details, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
...
    Page:
      type: object
      required: [items, count]
      properties:
        items: { type: array, items: {} }
        count: { type: integer, minimum: 0, description: Total matching items before page slicing }
        next_cursor: { type: string }
...
    Release:
      type: object
      required: [name, version, sha256, size, created_at]
      properties:
        name: { type: string }
        version: { type: string }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        size: { type: integer, format: int64, minimum: 0 }
        created_at: { type: string, format: date-time }
        signature: { type: string, format: byte }
        signature_status: { type: string, enum: [verified] }
        status: { type: string, enum: [published, pending_approval, approved] }
...
    DriftItem:
      type: object
      required: [device_name, package, kind, expected_sha256]
      properties:
        device_id: { type: string }
        device_name: { type: string }
        package: { type: string }
        kind: { type: string, enum: [missing, changed, extra] }
        expected_sha256: { type: string }
        observed_sha256: { type: string }
    PolicyDocument:
      type: object
      additionalProperties: false
      properties:
        require_signature: { type: boolean }
        require_approval: { type: boolean }
    PolicyMutation:
      oneOf:
        - { '$ref': '#/components/schemas/PolicyDocument' }
        - type: object
          additionalProperties: false
          required: [document]
          properties:
            document: { '$ref': '#/components/schemas/PolicyDocument' }
    Policy:
      type: object
      required: [id, workspace_id, revision, document, created_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        revision: { type: integer, format: int64 }
        document: { '$ref': '#/components/schemas/PolicyDocument' }
        created_at: { type: string, format: date-time }
...
    ReconcilePlan:
      type: object
      required: [device_id, workspace_id, manifest_revision, actions, generated_at]
      properties:
        device_id: { type: string }
        workspace_id: { type: string, format: uuid }
        manifest_revision: { type: integer, format: int64 }
        actions:
          type: array
          items:
            type: object
            required: [package, kind, to]
            properties:
              package: { type: string }
              kind: { type: string, enum: [install, update, remove] }
              from: { type: string }
              to: { type: string }
        generated_at: { type: string, format: date-time }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string, minLength: 1 }